move to rpc/layers and rename to auth client layer
